+++ src/router/e2fsprogs/lib/ext2fs/tst_libext2fs.c
@@ -38,7 +38,8 @@ static int print_blocks_proc(ext2_filsys fs EXT2FS_ATTR((unused)),
 			     void *private EXT2FS_ATTR((unused)))
 {
 	printf("%6lld %8llu (%d %llu)\n", (long long) blockcnt,
-	       (unsigned long long)*blocknr, ref_offset, ref_block);
+	       (unsigned long long) *blocknr, ref_offset,
+	       (unsigned long long) ref_block);
 	return 0;
 }
